# Define workloads specs by YAML