--- a/test/integration/scheduler_perf/config/node-default.yaml
+++ b/test/integration/scheduler_perf/config/node-default.yaml
@@ -1,0 +1,14 @@
+apiVersion: v1
+kind: Node
+metadata:
+  generateName: scheduler-perf-
+spec: {}
+status:
+  capacity:
+    pods: "110"
+    cpu: "4"
+    memory: 32Gi
+  conditions:
+    - status: "True"
+      type: Ready
+  phase: Running