--- a/test/integration/scheduler_perf/config/pv-aws.yaml
+++ b/test/integration/scheduler_perf/config/pv-aws.yaml
@@ -1,0 +1,10 @@
+apiVersion: v1
+kind: PersistentVolume
+spec:
+  accessModes:
+  - ReadOnlyMany
+  awsElasticBlockStore:
+    volumeID: <volume-id>
+  capacity:
+    storage: 1Gi
+  persistentVolumeReclaimPolicy: Retain