--- a/test/integration/scheduler_perf/config/pv-csi.yaml
+++ b/test/integration/scheduler_perf/config/pv-csi.yaml
@@ -1,0 +1,10 @@
+apiVersion: v1
+kind: PersistentVolume
+spec:
+  accessModes:
+  - ReadOnlyMany
+  capacity:
+    storage: 1Gi
+  csi:
+    driver: ebs.csi.aws.com
+  persistentVolumeReclaimPolicy: Retain