@liu-cong 2019-12-12
Generally LGTM now. If you can run the comparison with old benchmarks and verify the results, I will LGTM.