@notpad 2019-12-13
> Generally LGTM now. If you can run the comparison with old benchmarks and verify the results, I will LGTM.

Hi @liu-cong, I have two questions about how to compare the results:
1. share we run both tests with b.N or both without b.N ? 
2. what result should we pay attention to