@liu-cong 2019-12-13
> share we run both tests with b.N or both without b.N ?

I am not too sure about golang benchmark behavior if we don't use b.N. But what we want is to measure the average latency per pod to compare with old results. So if golang benchmark doesn't output that, you can simply create a timer and calculate yourself.

BTW, make sure you update to latest master and use the same number of nodes/pods with old tests. I think ahg recently updated the old benchmark tests to have only 1000/5000 nodes cases.

> what result should we pay attention to

We are interested in is the average latency to schedule a pod, or  how many `ns/op`.