@notpad 2019-12-18
Here is a comparison between the old benchmark test result and new performance test result:


TestName | Old B.N | Old Benchmark Result (ns/op) | New B.N | New Benchmark Result | (Old Benchmark Result * Old B.N) / (New Benchmark Result * New B.N)
-- | -- | -- | -- | -- | --
BenchmarkScheduling/500Nodes/500Pods-4 | 1000 | 3309331 | 1 | 3509292998 | 0.943019292
BenchmarkScheduling/5000Nodes/5000Pods-4 | 1000 | 4403971 | 1 | 4624369057 | 0.952339864
BenchmarkSchedulingPodAntiAffinity/500Nodes/100Pods-4 | 400 | 4328731 | 1 | 1965632244 | 0.880883189
BenchmarkSchedulingPodAntiAffinity/5000Nodes/1000Pods-4 | 1000 | 12640622 | 1 | 13511421154 | 0.935550884
BenchmarkSchedulingSecrets/500Nodes/500Pods-4 | 1000 | 2779263 | 1 | 2829301761 | 0.982314095
BenchmarkSchedulingSecrets/5000Nodes/5000Pods-4 | 1000 | 4474159 | 1 | 4895832208 | 0.913870985
BenchmarkSchedulingInTreePVs/500Nodes/500Pods-4 | 1000 | 9994053 | 1 | 9241005032 | 1.081489834
BenchmarkSchedulingInTreePVs/5000Nodes/5000Pods-4 | 1000 | 18669667 | 1 | 19642626271 | 0.950466946
BenchmarkSchedulingMigratedInTreePVs/500Nodes/500Pods-4 | 1000 | 8175321 | 1 | 8001034375 | 1.021783012
BenchmarkSchedulingMigratedInTreePVs/5000Nodes/5000Pods-4 | 1000 | 17471185 | 1 | 15534930510 | 1.124638761
BenchmarkSchedulingCSIPVs/500Nodes/500Pods-4 | 1000 | 8761479 | 1 | 7059429922 | 1.241102907
BenchmarkSchedulingCSIPVs/5000Nodes/5000Pods-4 | 1000 | 13196656 | 1 | 14505221562 | 0.909786586
BenchmarkSchedulingPodAffinity/500Nodes/500Pods-4 | 1000 | 6382893 | 1 | 5953229144 | 1.072173243
BenchmarkSchedulingPodAffinity/5000Nodes/5000Pods-4 | 1000 | 23111537 | 1 | 20988022216 | 1.10117746
BenchmarkSchedulingPreferredPodAffinity/500Nodes/500Pods-4 | 1000 | 5962353 | 1 | 5532426979 | 1.077710203
BenchmarkSchedulingPreferredPodAffinity/5000Nodes/5000Pods-4 | 1000 | 19507929 | 1 | 19564638698 | 0.997101419
BenchmarkSchedulingPreferredPodAntiAffinity/500Nodes/500Pods-4 | 1000 | 5055702 | 1 | 4644951222 | 1.088429514
BenchmarkSchedulingPreferredPodAntiAffinity/5000Nodes/5000Pods-4 | 1000 | 26074282 | 1 | 23037491310 | 1.131819505
BenchmarkSchedulingNodeAffinity/500Nodes/500Pods-4 | 1000 | 4056305 | 1 | 4040893632 | 1.003813851
BenchmarkSchedulingNodeAffinity/5000Nodes/5000Pods-4 | 1000 | 8379190 | 1 | 6927978325 | 1.209471163